@@ -1,34 +1,37 @@
 {
-  "name": "math-sum",
-  "version": "1.0.0",
-  "description": "Sum numbers",
-  "license": "MIT",
-  "repository": "sindresorhus/math-sum",
-  "author": {
-    "name": "Sindre Sorhus",
-    "email": "sindresorhus@gmail.com",
-    "url": "sindresorhus.com"
-  },
-  "engines": {
-    "node": ">=0.10.0"
-  },
-  "scripts": {
-    "test": "ava"
-  },
-  "files": [
-    "index.js"
-  ],
-  "keywords": [
-    "math",
-    "sum",
-    "number",
-    "numbers",
-    "num",
-    "add",
-    "count",
-    "array"
-  ],
-  "devDependencies": {
-    "ava": "*"
-  }
+	"name": "math-sum",
+	"version": "1.0.0",
+	"description": "Sum numbers",
+	"license": "MIT",
+	"repository": "sindresorhus/math-sum",
+	"funding": "https://github.com/sponsors/sindresorhus",
+	"author": {
+		"name": "Sindre Sorhus",
+		"email": "sindresorhus@gmail.com",
+		"url": "https://sindresorhus.com"
+	},
+	"type": "module",
+	"exports": "./index.js",
+	"engines": {
+		"node": ">=12"
+	},
+	"scripts": {
+		"test": "xo && ava"
+	},
+	"files": [
+		"index.js"
+	],
+	"keywords": [
+		"math",
+		"sum",
+		"number",
+		"numbers",
+		"add",
+		"count",
+		"array"
+	],
+	"devDependencies": {
+		"ava": "^3.15.0",
+		"xo": "^0.38.2"
+	}
 }
