@@ -1,8 +1,8 @@
 import test from 'ava';
-import fn from './';
+import mathSum from './index.js';
 
-test(t => {
-	t.is(fn(1, 2, 3, 4), 10);
-	t.is(fn(1000, 1000), 2000);
-	t.is(fn([1, 2, 3, 4]), 10);
+test('main', t => {
+	t.is(mathSum(1, 2, 3, 4), 10);
+	t.is(mathSum(1000, 1000), 2000);
+	t.is(mathSum([1, 2, 3, 4]), 10);
 });
